bump ktmidi version and fix build to match the latest API.

+++ mugene/src/commonMain/kotlin/dev/atsushieno/mugene/mml_smf_generator.kt
@@ -14,7 +14,7 @@ class MmlSmfGenerator(private val source: MmlResolvedMusic) {
         }
     }
 
-    var result: MidiMusic = MidiMusic().apply { deltaTimeSpec = (source.baseCount / 4).toByte() }
+    var result: MidiMusic = MidiMusic().apply { deltaTimeSpec = source.baseCount / 4 }
 
     private fun generateSong() {
         for (t in source.tracks)
